feat(dilesa-ventas-expediente-tabs): Sprint 1 — routed tabs del expediente (persisten en captura)
El detalle de venta deja de ser un page monolítico de 2,642 L con tabs en
useState y pasa a routed tabs montados en un layout compartido:

- components/dilesa/venta-detalle/{types,provider,ui,shell}.tsx: el cerebro
  (VentaDetalleProvider + useVentaDetalle), sub-componentes y Shell.
- app/dilesa/ventas/[id]/layout.tsx: rama expediente (provider + shell + tabs)
  y rama captura (persiste la barra de tabs encima del CapturarFaseHeader) →
  el menú persiste al capturar una fase.
- 4 páginas de tab (Operación / Cuadratura / Documentos / Bitácora), cada una
  con su <RequireAccess> fino. Pipeline y Estado de cuenta siguen en Operación
  (salen a tab propio en S2).

RBAC: sub-slugs dilesa.ventas.{operacion,cuadratura,documentos,bitacora} en
ROUTE_TO_MODULE + EXPECTED_DB_MODULE_SLUGS + MODULE_DEPS (cadena
lista->operacion->{capturas,tabs}) + migración con backfill de permisos
(pendiente de aplicar por Beto).
